+++ common/src/main/java/foundry/imgui/impl/ImGuiMCMixinPlugin.java
@@ -32,7 +32,8 @@ public abstract class ImGuiMCMixinPlugin implements IMixinConfigPlugin {
     public List<String> getMixins() {
         //? if >=1.21.6 {
         /*return List.of("renderer.v1.AbstractTextureMixin", "renderer.v1.GpuSamplerMixin", "renderer.v1.GpuTextureViewMixin");
-         *///? } else
-        return List.of();
+         *///? } else {
+        return List.of("renderer.v0.AbstractTextureMixin");
+        //? }
     }
 }