+++ common/src/main/java/foundry/imgui/mixin/renderer/v0/AbstractTextureMixin.java
@@ -1,9 +1,7 @@
-package foundry.imgui.mixin.renderer.v1;
+package foundry.imgui.mixin.renderer.v0;
 
-//? if >=1.21.6 {
+//? if <1.21.6 {
 
-/*import com.mojang.blaze3d.textures.GpuTextureView;
-import foundry.imgui.api.ImGuiMC;
 import foundry.imgui.api.ImGuiTextureProvider;
 import net.minecraft.client.renderer.texture.AbstractTexture;
 import org.spongepowered.asm.mixin.Mixin;
@@ -13,17 +11,16 @@ import org.spongepowered.asm.mixin.Shadow;
 public abstract class AbstractTextureMixin implements ImGuiTextureProvider {
 
     @Shadow
-    public abstract GpuTextureView getTextureView();
+    public abstract int getId();
 
     @Override
     public long imguimc$id() {
-        return ImGuiMC.getTexture(this.getTextureView()).imguimc$id();
+        return this.getId();
     }
 
     @Override
     public void imguimc$setId(final long id) {
-        ImGuiMC.getTexture(this.getTextureView()).imguimc$setId(id);
     }
 }
 
-*///?}
+//?}