+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 - Add `ImGuiLoadEvent` that fires after ImGui has successfully loaded
+- Remove interface injection
+- Added getters for `ImGuiTextureProvider` and `ImGuiSampler`